refactor: dto 불필요한 정적 변수 제거
