+++ src/main/java/woowacourse/shoppingcart/dto/CustomerResponse.java
@@ -1,6 +1,5 @@
 package woowacourse.shoppingcart.dto;
 
-import woowacourse.auth.dto.PhoneNumber;
 import woowacourse.shoppingcart.domain.Customer;
 
 public class CustomerResponse {
@@ -22,7 +21,8 @@ public class CustomerResponse {
 
     public static CustomerResponse of(Customer customer) {
         final PhoneNumber phoneNumber = PhoneNumber.of(customer.getPhoneNumber());
-        return new CustomerResponse(customer.getId(), customer.getAccount(), customer.getNickname(), customer.getAddress(), phoneNumber);
+        return new CustomerResponse(customer.getId(), customer.getAccount(), customer.getNickname(),
+                customer.getAddress(), phoneNumber);
     }
 
     public long getId() {
