+++ src/main/java/woowacourse/shoppingcart/dto/PhoneNumber.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.shoppingcart.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
