+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import woowacourse.auth.dto.DeleteCustomerRequest;
-import woowacourse.auth.dto.PhoneNumber;
-import woowacourse.auth.dto.UpdateCustomerRequest;
+import woowacourse.shoppingcart.dto.DeleteCustomerRequest;
+import woowacourse.shoppingcart.dto.PhoneNumber;
+import woowacourse.shoppingcart.dto.UpdateCustomerRequest;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.CustomerResponse;
